+++ examples/experimental/nodes/frontend/next.config.js
@@ -1,0 +1,18 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+    output: 'standalone',
+    webpack: (config, { isServer }) => {
+      if (!isServer) {
+        config.resolve.fallback = {
+          ...config.resolve.fallback,
+          net: false,
+          tls: false,
+          fs: false
+        };
+      }
+      return config;
+    },
+    reactStrictMode: true,
+  };
+  
+  module.exports = nextConfig;
