--- examples/experimental/nodes/frontend/src/next.config.js
@@ -1,16 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-    webpack: (config, { isServer }) => {
-      if (!isServer) {
-        config.resolve.fallback = {
-          ...config.resolve.fallback,
-          net: false,
-          tls: false,
-          fs: false
-        };
-      }
-      return config;
-    }
-  };
-  
-  module.exports = nextConfig;
